IR-LSQ refactor: sparse-input benchmark, fix breakdown + analytical_kb bugs
- Replace NMR/Kronecker benchmark with CQRRT_linop_irlsq.cc.
  Loads a tall sparse .mtx, wraps as SparseLinOp, generates synthetic
  x_true + b = J*x_true + noise. For each Q-less QR variant: draws a
  fresh sparse sketch S2 (independent of CQRRT's S1), forms x_0 =
  R^{-1} R^{-T} (S2 A)^T (S2 b) (paper Algorithm 1, line 3, Q-less
  form), then runs 2-step IR with inner CG.

- Strip Tikhonov from IterRefineLSQ. Drop KroneckerOperator,
  RegularizedLinOp, and their tests/includes/CMake entries. Drop the
  GEQP3-stabilized variant from the benchmark and rename
  CQRRT_linop_stb_bqrrp -> CQRRT_linop_bqrrp.

- Fix IR-LSQ populate_times double-counting bug. Inner CG's
  TRSM/fwd/adj contributions were tracked into both t_inner_total
  (wallclock of the inner_cg call) and the shared
  t_{trsm,fwd,adj}_total counters, so 'other = outer_total - inner -
  trsm - fwd - adj' went negative. Split into outer-only and
  inner-only counters and report inner_ex = t_inner_total -
  t_inner_{trsm,fwd,adj}, total_trsm = t_outer_trsm + t_inner_trsm
  (similarly fwd, adj). 'other' is now a non-negative residue
  (axpy/copy/nrm2 bookkeeping).

- Fix three analytical_kb formulas in rl_memory_tracker.hh.
  cqrrt_linops_bqrrp_analytical_kb captured only the BQRRP-precond
  moment (d*n + 4n^2) and missed the later Gram-loop moment (d*n + n
  + n^2 + m*b_eff). For tall inputs the latter dominates. Now
  returns the max of the two; signature extended to (m, n, d_factor,
  block_size). scholqr3_linops_analytical_kb forgot G3_factor at the
  iter-3 peak (5n^2 -> 6n^2). scholqr3_linops_basic_analytical_kb
  forgot all three G_i_factor members (3n^2 -> 6n^2).
