+++ RandLAPACK/linops/rl_linops.hh
@@ -16,5 +16,3 @@
 #include "rl_composite_linop.hh"
 #include "rl_sym_linops.hh"
 #include "rl_materialize.hh"
-#include "rl_kronecker_linop.hh"
-#include "rl_regularized_linop.hh"
