+++ benchmark/CMakeLists.txt
@@ -159,8 +159,8 @@ set(
 add_benchmark(NAME CQRRT_linop_applications           CXX_SOURCES bench_CQRRT_linops/CQRRT_linop_applications.cc           LINK_LIBS ${Benchmark_libs_cqrrt_linops})
 add_benchmark(NAME CQRRT_linop_basic                  CXX_SOURCES bench_CQRRT_linops/CQRRT_linop_basic.cc                  LINK_LIBS ${Benchmark_libs_cqrrt_linops})
 add_benchmark(NAME CQRRT_diagnostic                   CXX_SOURCES bench_CQRRT_linops/CQRRT_diagnostic.cc                   LINK_LIBS ${Benchmark_libs_cqrrt_linops})
-# NMR / Kronecker-operator benchmark — no Eigen / fast_matrix_market dependency
-add_benchmark(NAME CQRRT_linop_nmr                    CXX_SOURCES bench_CQRRT_linops/CQRRT_linop_nmr.cc                    LINK_LIBS ${Benchmark_libs})
+# Sparse-input IR-LSQ benchmark: Q-less QR + sketch-and-solve x0 + 2-step IR.
+add_benchmark(NAME CQRRT_linop_irlsq                  CXX_SOURCES bench_CQRRT_linops/CQRRT_linop_irlsq.cc                  LINK_LIBS ${Benchmark_libs_cqrrt_linops})
 
 # ABRIK benchmarks
 add_benchmark(NAME ABRIK_runtime_breakdown        CXX_SOURCES bench_ABRIK/ABRIK_runtime_breakdown.cc        LINK_LIBS ${Benchmark_libs})
