+++ test/drivers/test_iter_refine_lsq.cc
@@ -3,10 +3,8 @@
 //
 // Strategy: wrap a small dense tall A as a DenseLinOp, build R from
 // lapack::geqrf on a copy of A, then verify the IR-LSQ solution matches
-// a closed-form reference:
-//   * unregularized cases  → lapack::gels on (A, b)
-//   * Tikhonov case        → lapack::posv on (A^T A + lambda^2 I, A^T b)
-// across well-conditioned, residualful, regularized, and imperfect-R cases.
+// the closed-form lapack::gels reference across well-conditioned,
+// residualful, and imperfect-R cases.
 
 #include <RandBLAS.hh>
 #include <RandLAPACK.hh>
@@ -125,51 +123,6 @@ TEST_F(TestIterRefineLSQ, dense_with_residual) {
     int status = ir.call(J, R.data(), n, b.data(), m, x_ir.data(), n);
     EXPECT_EQ(status, 0);
 
-    T diff_norm = 0;
-    for (int64_t i = 0; i < n; ++i) {
-        T d = x_ir[i] - x_ref[i];
-        diff_norm += d * d;
-    }
-    diff_norm = std::sqrt(diff_norm);
-    T xref_norm = blas::nrm2(n, x_ref.data(), 1);
-    EXPECT_LT(diff_norm / xref_norm, 1e-10);
-}
-
-
-// Tikhonov regularization: solving min ||Jx-b||² + λ²||x||² should match
-// the closed-form solution x = (J^T J + λ² I)^{-1} J^T b.
-TEST_F(TestIterRefineLSQ, tikhonov_regularization) {
-    using T = double;
-    int64_t m = 80, n = 12;
-    T lambda = 0.3;
-
-    std::vector<T> A(m * n), b(m);
-    fill_random(A, 101);
-    fill_random(b, 102);
-
-    // Reference: solve (A^T A + λ² I) x = A^T b directly via lapack::posv.
-    std::vector<T> Gram(n * n, 0.0), AtB(n, 0.0);
-    blas::syrk(blas::Layout::ColMajor, blas::Uplo::Upper, blas::Op::Trans,
-               n, m, (T)1.0, A.data(), m, (T)0.0, Gram.data(), n);
-    for (int64_t i = 0; i < n; ++i) Gram[i + i * n] += lambda * lambda;
-    blas::gemv(blas::Layout::ColMajor, blas::Op::Trans, m, n, (T)1.0, A.data(), m,
-               b.data(), 1, (T)0.0, AtB.data(), 1);
-    std::vector<T> x_ref(AtB);
-    lapack::posv(blas::Uplo::Upper, n, 1, Gram.data(), n, x_ref.data(), n);
-
-    // Build R from QR(A) — perfect preconditioner for J^T J alone, plus the
-    // Tikhonov shift gets added inside the inner CG.
-    std::vector<T> R(n * n, 0);
-    build_R_from_A(A.data(), m, n, R.data(), n);
-
-    DenseLinOp<T> J(m, n, A.data(), m, Layout::ColMajor);
-    IterRefineLSQ<T> ir(/*tol=*/1e-12, /*max_inner=*/200, /*n_steps=*/2,
-                       /*timing=*/false, /*verbose=*/false, /*lambda=*/lambda);
-    std::vector<T> x_ir(n, 0);
-    int status = ir.call(J, R.data(), n, b.data(), m, x_ir.data(), n);
-    EXPECT_EQ(status, 0);
-
-    // Compare: x_ir vs x_ref, both solving the same regularized normal equations.
     T diff_norm = 0;
     for (int64_t i = 0; i < n; ++i) {
         T d = x_ir[i] - x_ref[i];
